refactor(plugin-image): pixabay search component split
